# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Avito Fair Queue API
  version: 1.0.0
  description: >-
    API MVP «Авито Честная очередь». Сервер принимает решения о FIFO,
    резервировании и доступе к checkout на основе PostgreSQL.
servers:
  - url: http://localhost:8080
    description: Local Dev Server
...
paths:
...
  /api/v1/products/{product_id}/queue/join:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    post:
      tags: [Queue]
      summary: Идемпотентно войти в очередь
      description: >-
        Пользователь определяется только заголовком X-Demo-User-ID. Тело
        запроса не содержит user_id.
      operationId: joinQueue
      security:
        - DemoUser: []
      responses:
        '200':
          description: Возвращена текущая или новая активная заявка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ProductNotFound'
        '409':
          $ref: '#/components/responses/QueueConflict'
  /api/v1/products/{product_id}/queue/me:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      tags: [Queue]
      summary: Получить актуальное состояние своей заявки
      operationId: getMyQueueState
      security:
        - DemoUser: []
      responses:
        '200':
          description: Состояние заявки получено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/QueueEntryNotFound'
    delete:
      tags: [Queue]
      summary: Выйти из очереди или отказаться от права
      operationId: leaveQueue
      security:
        - DemoUser: []
      responses:
        '200':
          description: Состояние заявки обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/QueueEntryNotFound'
        '409':
          $ref: '#/components/responses/QueueConflict'
...
components:
  securitySchemes:
    DemoUser:
      type: apiKey
      in: header
      name: X-Demo-User-ID
      description: Идентичность пользователя из доверенного middleware через `X-Demo-User-ID`.
  parameters:
    ProductId:
      name: product_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: Demo-идентичность отсутствует или недействительна.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ProductNotFound:
      description: Товар не найден.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    QueueEntryNotFound:
      description: Заявка пользователя на этот товар не найдена.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    QueueConflict:
      description: Операция очереди невозможна в текущем состоянии.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    QueueState:
      type: object
      additionalProperties: false
      required:
        - queue_entry_id
        - product_id
        - ticket_no
        - status
        - message
        - next_action
        - grant
      properties:
        queue_entry_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        ticket_no:
          type: integer
          format: int64
          minimum: 1
        status:
          $ref: '#/components/schemas/QueueEntryStatus'
        position:
          type: [integer, 'null']
          minimum: 1
          description: Позиция среди WAITING-заявок; не вычисляется клиентом.
        message:
          type: string
          description: Русскоязычное объяснение текущего состояния.
        next_action:
          type: string
          description: Русскоязычное действие, доступное пользователю.
        grant:
          anyOf:
            - $ref: '#/components/schemas/Grant'
            - type: 'null'
    QueueEntryStatus:
      type: string
      enum:
        - JOINING
        - WAITING
        - GRANTED
        - CHECKOUT_PENDING
        - PURCHASED
        - EXPIRED
        - PAYMENT_FAILED
        - SOLD_OUT
        - CANCELLED
        - ERROR
    Grant:
      type: object
      additionalProperties: false
      required:
        - id
        - product_id
        - inventory_unit_id
        - status
        - expires_at
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        inventory_unit_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/GrantStatus'
        expires_at:
          type: string
          format: date-time
    GrantStatus:
      type: string
      enum: [ACTIVE, CHECKOUT_PENDING, PURCHASED, EXPIRED, FAILED, CANCELLED]
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    ErrorBody:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code:
          type: string
          enum:
            - INVALID_REQUEST
            - UNAUTHORIZED
            - NOT_FOUND
            - NOT_IMPLEMENTED
            - METHOD_NOT_ALLOWED
            - INTERNAL
            - PRODUCT_NOT_FOUND
            - QUEUE_REQUIRED
            - QUEUE_ENTRY_NOT_FOUND
            - GRANT_NOT_FOUND
            - GRANT_EXPIRED
            - GRANT_ALREADY_USED
            - PRODUCT_SOLD_OUT
            - INVALID_STATE
            - SERVICE_UNAVAILABLE
        message:
          type: string
        request_id:
          type: string
          minLength: 1